NeRF updates and add inference codes - Add repeat to dataset loader to antisipate in consistant batch - Add more camera utils - Modify nerf callback to include last train image prediction - Modify nerf callback to inclode model saving - Add model path parameter on NeRF to load the trained model

+++ keras_nerf/data/rays.py
@@ -118,8 +118,9 @@ class RaysGenerator:
         # Add noise to the sample points
 
         noise = (tf.random.uniform(shape=[self.image_width,
-                 self.image_height, self.n_sample], dtype=tf.float32) *
+                 self.image_height, self.n_sample]) *
                  (self.far - self.near) / self.n_sample_fl)
+        # tf.print(noise)
 
         sample_points = sample_points + noise  # Shape: [H, W, N]
